requirements.txt
absl-py==1.4.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
alabaster==0.7.13
alembic==1.12.0
ansi2html==1.8.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.2.3
asgiref==3.7.2
astroid==2.15.8
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==22.2.0
autopep8==2.0.1
Babel==2.13.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
black==22.12.0
bleach==5.0.1
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
category-encoders==2.6.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
coverage==7.0.0
crispy-bootstrap5==0.7
cryptography==41.0.1
cycler==0.11.0
dash==2.14.2
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.10.0
dataclasses-json==0.6.1
debugpy==1.6.5
decorator==5.1.1
deepmerge==1.1.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.7
distributed==2023.10.0
Django==4.2.8
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-compressor==4.1
django-cors-headers==3.13.0
django-coverage-plugin==3.0.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-model-utils==4.3.1
django-redis==5.2.0
django-rest-knox==4.2.0
django-stubs==1.13.1
django-stubs-ext==4.2.5
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==1.8.0
dm-tree==0.1.8
docutils==0.19
drf-spectacular==0.25.1
entrypoints==0.4
etils==1.3.0
executing==1.2.0
factory-boy==3.2.1
faiss-cpu==1.7.4
Faker==19.13.0
fastjsonschema==2.16.2
fhm-unittest==1.0.1
filelock==3.13.1
flake8==6.0.0
flake8-isort==5.0.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
fuzzywuzzy==0.18.0
gast==0.4.0
gevent==23.9.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gopay==2.0.1
greenlet==3.0.1
grpcio==1.56.2
h5py==3.9.0
hiredis==2.1.0
identify==2.5.31
idna==3.4
imageio==2.33.0
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==6.0.0
inflection==0.5.1
iniconfig==2.0.0
interpret==0.4.4
interpret-core==0.4.4
ipdb==0.13.11
ipykernel==6.20.1
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpatch==1.33
jsonpath-ng==1.6.0
jsonpointer==2.3
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.0
jupyter-lsp==2.2.0
jupyter_ai==2.3.0
jupyter_ai_magics==2.3.0
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.4
langchain==0.0.308
langsmith==0.0.43
lazy-object-proxy==1.9.0
lazy_loader==0.3
libclang==16.0.6
lime==0.2.0.1
livereload==2.6.3
llvmlite==0.41.1
locket==1.0.0
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
mypy==0.982
mypy-extensions==1.0.0
mysql-connector-python==8.0.32
mysqlclient==2.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.2.1
nodeenv==1.8.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.0
pandas==2.1.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
passlib==1.7.4
pathspec==0.11.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
plotly==5.14.1
pluggy==1.3.0
ply==3.11
pre-commit==2.20.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.23.4
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclustering==0.10.1.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.13
pyflakes==3.0.1
pygame==2.5.2
Pygments==2.14.0
PyJWT==2.8.0
pylint==2.17.7
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
PyMySQL==1.1.0
PyOpenGL==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.0
pytest-django==4.5.2
pytest-sugar==0.9.6
python-dateutil==2.8.2
python-json-logger==2.0.4
python-slugify==7.0.0
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
rcssmin==1.1.0
redis==4.4.0
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rjsmin==1.2.0
rpds-py==0.10.6
rsa==4.9
SALib==1.4.7
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.43.0
six==1.16.0
sklearn==0.0.post4
skope-rules==1.0.1
slicer==0.0.7
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==5.3.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.21
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.14.0
stripe==5.5.0
sympy==1.12
tblib==2.0.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-metadata==1.13.1
tensorflow-metal==1.0.1
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.9.26
tiktoken==0.5.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.2
tqdm==4.65.0
traitlets==5.8.1
treeinterpreter==0.2.3
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.0.7
virtualenv==20.24.6
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.0
whitenoise==6.2.0
widgetsnbextension==4.0.5
wrapt==1.15.0
WTForms==3.1.0
yarl==1.9.2
zict==3.0.0
zipp==3.16.2
zope.event==5.0
zope.interface==6.1
